Revisions after PR review
- Rename GraphPos.node to block for clarity
- Simplify matching logic by removing redundant normalization helpers
- Use thiserror for search index error types
- Move tests inline, delete separate test file
- Add Workspace helpers for search index directory management
- Clean up main.rs unused methods and formatting
